feat(api): 오더 API 기본 기능 구현

+++ src/services/order-service.js
@@ -2,36 +2,39 @@ import { OrderModel } from '../db';
 import { AppError, commonErrors } from '../middlewares';
 
 class OrderService {
-  async createOrder(orderInfo) {
+  static async createOrder(orderInfo) {
     const newOrder = await OrderModel.create(orderInfo);
     return newOrder;
   }
 
-  async getMyOrders(userId) {
+  static async getMyOrders(userId) {
     const orders = await OrderModel.findAllByUserId(userId);
     return orders;
   }
+  static async getOrderById(orderId) {
+    const order = await OrderModel.findById(orderId);
+    return order;
+  }
 
-  async getOrders() {
+  static async getOrders() {
     const orders = await OrderModel.findAll();
     return orders;
   }
 
-  async getOrder(orderId) {
+  static async getOrder(orderId) {
     const order = await OrderModel.findById(orderId);
     if (!order) {
       throw new Error('해당하는 주문 정보가 없습니다. 다시 확인해주세요!');
     }
     return order;
   }
-
-  async updateOrder(orderId, updateInfo) {
+  static async updateOrder(orderId, updateInfo) {
     const order = await OrderModel.findById(orderId);
     if (order.status !== '배송 준비 중') {
       throw new AppError(
         commonErrors.businessError,
         400,
-        '배송이 시작된 상품은 취소할 수 없어요 :(',
+        '배송이 시작된 상품은 수정/취소할 수 없어요 :(',
       );
     }
     const updatedOrder = await OrderModel.update({
@@ -45,7 +48,11 @@ class OrderService {
     const order = await OrderModel.findById(orderId);
     if (!order) {
       if (!order) {
-        throw new Error('해당하는 주문 정보가 없습니다. 다시 확인해주세요!');
+        throw new AppError(
+          commonErrors.inputError,
+          400,
+          '해당하는 주문 정보가 없습니다. 다시 확인해주세요!',
+        );
       }
     }
     const updateInfo = { deletedAt: Date.now() };
@@ -53,7 +60,15 @@ class OrderService {
     return result;
   }
 
-  async deleteOrder(orderId) {
+  static async deleteOrder(orderId) {
+    const order = await OrderModel.findById(orderId);
+    if (!order.deletedAt) {
+      throw new AppError(
+        commonErrors.businessError,
+        400,
+        '회원이 삭제하지 않은 주문 내역은 삭제할 수 없습니다.',
+      );
+    }
     const result = await OrderModel.delete(orderId);
     return result;
   }
